backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.8.2
chromadb==0.5.5
pypdf==5.0.1
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.4.0
rapidfuzz==3.9.6
pydantic-settings==2.5.2
google-genai
python-docx>=0.8.11
sqlalchemy==2.0.31
python-jose==3.3.0
authlib==1.3.1
passlib==1.7.4
requests==2.31.0
